tempcal: change to double as float was insufficient for calculation

+++ src/modules/tempcal/matrix_alg.h
@@ -47,5 +47,5 @@
 
 float *mat_mul(float *A, float *B, uint8_t n);
 bool mat_inverse(float *A, float *inv, uint8_t n);
-bool inverse4x4(float m[], float invOut[]);
+bool inverse4x4(double m[], double invOut[]);
 bool inverse(float *A, float *inv, uint8_t n);
